requirements.txt
torch
numpy
requests
tqdm
h5py
scikit-image
pyyaml
opencv-python
matplotlib
wandb
hydra-core==1.3.*
torchmetrics
pytorch-lamb
torch-optimizer
sophia-opt
lion-pytorch
adabelief-pytorch
pytorch-msssim
pandas
deepspeed
mpi4py